feat(WGConfValidator): add new class for validation
The WGConfValidator class currently only has methods to check key
validity. It will gain more methods with time.

Note that it's important to keep an eye on this class. I can see it
becoming too big and unwieldy. If it approaches that point, it's
probably best to split some of the functionality into other classes.