feat(validation): two new functions
This continues the same effort of moving things out of main into
dedicated functions that should act like miniature programs.
validate_ips is kept as a separate function despite being similar to
extract_ips because their roles are different.

validate_wg_data is added to decouple validation of the data and
printing + formatting.