feat(convert_wg_to_hostname_if): new function
Instead of printing a bunch of statements based on conditionals and
having a mess in the main code, put it in a function that returns a list
of strings. print() will add a newline to each of them in a for loop, so
it works out well.

I don't think in its current form that the function is perfect. It
still messes with the data, which I don't like. But for now it gives me
a point to work from in the future.