fix(default_routes): restrict to first of ip4/ip6
OpenBSD only supports one default route for IPv4 and IPv6 as far
as I know. I'm unsure how I feel about this solution, but it works
for now. It makes me wonder if the IP addresses for the interface
should be restricted to the first one for IPv4 and IPv6 as well. I've
left them as is for now.

fix(convert_wg_to_hostname_if): zap trailing space