fix(extract_ips): add ValueError except back in
I do need it after all because if a list contained "invalid_ip" (a
string), a more useful behavior for extract_ips is to continue on
rather than raising an exception and stopping the processing there.
I'm trying out pytest out-of-tree (will commit sooner or later) and
it caught this.