fix(transform_wg_data): better type hinting
Note that fully hinting the return value is a huge pain. It's
possible to fully hint it, but then other parts of the code will
cause type checker complaints because str does not have the ip or
version methods like ipaddress.IPv4Interface and ipaddress.IPv6Interface
do. To fix this, the data structure and function would need to be
rearranged so that these are separate:

- dict[str, str]
- dict[str, list[ipaddress.IPv4Interface, ipaddress.IPv6Interface]

This seems horribly tedious for no obvious gain, so don't do it
right now.