test(parse_cli_arguments): remove asserts
The exit code really isn't that important in the grand scheme of things,
and this doesn't actually seem to get executed anyway based on code
coverage.

The SystemExit exception in the main code probably should be replaced
with something more appropriate at some point, and then these tests
would need to be updated. But for now it stays like this.